client/requirements.txt
flask
flask-cors
requests
pycryptodome
redis
